requirements.txt
Flask==3.1.0
pytest==8.3.5
PyYAML==6.0.2
types-PyYAML==6.0.12.20250402
Requests==2.32.3
types-requests==2.32.0.20250328
selenium==4.32.0
termcolor==3.1.0
Werkzeug==3.1.3
